Déplacement du composant GeneratorAlbum, correction des imports, refactorisation et correction de l'indentation. Il manquera le refacto de son CSS qui est en px et non en rem
