+++ Avancement.txt
@@ -39,7 +39,8 @@ Note d'information, dans la navigation du site,
 Les page compte et panier vont être retirer provisoirement le temps de les éditer.
 
 ------------------------------------------------------------------------------
-
+Refactorisation du CSS : 
+- Dans le dossier Elements
 ------------------------------------------------------------------------------
 
 Il faudra créer de nouveau hook pour les fetchs
